Function check_cpu::do_comparison

fn do_comparison(
    args: &Args, 
    start: &Calculations, 
    end: &Calculations
) -> Status

Check if we have exceeded the warn or critical limits

The *cstat parameters, if present, mean that in addition to performing the standard checks we check the currently-running container against the same thresholds.